js: Promote mock provisioner to package util test
